times_database/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring & Logging
structlog==23.2.0



